requirements.txt
requests>=2.28.0
supabase>=2.0.0
python-dotenv>=1.0.0
Pillow>=9.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.21.0
clip
numpy>=1.21.0
pandas>=1.5.0
tqdm>=4.60.0
aiohttp>=3.8.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
fake-useragent>=1.0.0
